refactor(host): extract Beads lifecycle subsystem
Move repo readiness, verification, and provisioning behind a dedicated lifecycle subsystem so BeadsTaskStore stays focused on persistence. This keeps host-owned Beads and shared-Dolt behavior easier to test and extend without growing task CRUD code.
