+++ apps/desktop/src-tauri/crates/host-infra-beads/src/lifecycle/mod.rs
@@ -1,0 +1,6 @@
+mod context;
+mod coordinator;
+mod provisioner;
+mod verifier;
+
+pub(crate) use coordinator::BeadsLifecycle;
