+++ apps/desktop/src-tauri/crates/host-infra-beads/src/store/task_ops.rs
@@ -1,20 +1,9 @@
 use super::*;
 use chrono::{Duration as ChronoDuration, Utc};
 use serde::Deserialize;
+use std::collections::HashSet;
 
 impl BeadsTaskStore {
-    fn reason_requires_shared_database_seed(reason: &str) -> bool {
-        // `bd where --json` currently reports missing shared databases via free-form
-        // error strings instead of a structured machine-readable code. Match only the
-        // exact server-missing variants we have observed so other verification failures
-        // still follow the normal repair/init path.
-        let normalized = reason.to_ascii_lowercase();
-        normalized.contains("not found on dolt server")
-            || normalized.contains("server not reachable")
-            || normalized.contains("dolt server unreachable")
-            || normalized.contains("error 1049")
-    }
-
     fn append_raw_issue_list(
         &self,
         value: serde_json::Value,
@@ -60,179 +49,8 @@ impl BeadsTaskStore {
         }
     }
 
-    fn beads_store_footprint_exists(beads_dir: &Path) -> bool {
-        beads_dir.join("metadata.json").exists() || beads_dir.join("beads.db").exists()
-    }
-
-    fn task_status_requires_custom_configuration(status: &TaskStatus) -> bool {
-        matches!(
-            status,
-            TaskStatus::SpecReady
-                | TaskStatus::ReadyForDev
-                | TaskStatus::AiReview
-                | TaskStatus::HumanReview
-        )
-    }
-
-    fn ensure_existing_store_is_ready(
-        &self,
-        repo_path: &Path,
-        beads_dir: &Path,
-        reason: &str,
-    ) -> Result<()> {
-        if Self::reason_requires_shared_database_seed(reason) {
-            self.materialize_shared_database_from_attachment(repo_path, beads_dir)?;
-        } else {
-            self.repair_repo_store(repo_path)?;
-        }
-
-        let (is_ready_after_repair, reason_after_repair) =
-            self.verify_repo_initialized(repo_path, beads_dir)?;
-        if !is_ready_after_repair {
-            self.ensure_new_store_is_ready(repo_path, beads_dir, &reason_after_repair)?;
-        }
-        Ok(())
-    }
-
-    fn materialize_shared_database_from_attachment(
-        &self,
-        repo_path: &Path,
-        beads_dir: &Path,
-    ) -> Result<()> {
-        let backup_dir = beads_dir.join("backup");
-        if !backup_dir.is_dir() {
-            return Err(anyhow!(
-                "Shared Dolt database is missing for {} and no attachment backup exists at {}",
-                beads_dir.display(),
-                backup_dir.display()
-            ));
-        }
-
-        let database_name = compute_beads_database_name(repo_path)?;
-        if self.command_runner.uses_real_processes() {
-            restore_shared_dolt_database_from_backup(
-                std::process::id(),
-                database_name.as_str(),
-                &backup_dir,
-            )?;
-        } else {
-            let shared_dolt_root = resolve_shared_dolt_root()?;
-            let backup_url = format!("file://{}", backup_dir.display());
-            self.command_runner.run_with_env(
-                "dolt",
-                &[
-                    "backup",
-                    "restore",
-                    backup_url.as_str(),
-                    database_name.as_str(),
-                ],
-                Some(&shared_dolt_root),
-                &[],
-            )?;
-        }
-
-        Ok(())
-    }
-
-    fn ensure_new_store_is_ready(
-        &self,
-        repo_path: &Path,
-        beads_dir: &Path,
-        init_failure_reason: &str,
-    ) -> Result<()> {
-        let slug = compute_repo_slug(repo_path);
-        let database_name = compute_beads_database_name(repo_path)?;
-        let env = self.build_bd_env(repo_path)?;
-        let env_refs = env
-            .iter()
-            .map(|(key, value)| (key.as_str(), value.as_str()))
-            .collect::<Vec<_>>();
-        let server_port = env
-            .iter()
-            .find_map(|(key, value)| (key == "BEADS_DOLT_SERVER_PORT").then_some(value.as_str()))
-            .ok_or_else(|| anyhow!("Missing BEADS_DOLT_SERVER_PORT while initializing repo"))?;
-        let working_dir = self.ensure_beads_working_dir(repo_path)?;
-        let (ok, _stdout, stderr) = self.command_runner.run_allow_failure_with_env(
-            "bd",
-            &[
-                "init",
-                "--server",
-                "--server-host",
-                "127.0.0.1",
-                "--server-port",
-                server_port,
-                "--server-user",
-                "root",
-                "--quiet",
-                "--skip-hooks",
-                "--skip-agents",
-                "--prefix",
-                slug.as_str(),
-                "--database",
-                database_name.as_str(),
-            ],
-            Some(&working_dir),
-            &env_refs,
-        )?;
-
-        if !ok {
-            let details = if stderr.trim().is_empty() {
-                init_failure_reason.to_string()
-            } else {
-                stderr.trim().to_string()
-            };
-            return Err(anyhow!(
-                "Failed to initialize Beads at {}: {}",
-                beads_dir.display(),
-                details
-            ));
-        }
-
-        let (is_ready_after_init, reason_after_init) =
-            self.verify_repo_initialized(repo_path, beads_dir)?;
-        if !is_ready_after_init {
-            return Err(anyhow!(
-                "Beads init completed but store is not ready at {}: {}",
-                beads_dir.display(),
-                reason_after_init
-            ));
-        }
-
-        Ok(())
-    }
-
     pub(super) fn ensure_repo_initialized_impl(&self, repo_path: &Path) -> Result<()> {
-        let repo_key = Self::repo_key(repo_path);
-        let lock = self.repo_lock(&repo_key)?;
-        let _guard = lock
-            .lock()
-            .map_err(|_| anyhow!("Beads repo lock poisoned"))?;
-
-        let beads_dir = resolve_repo_beads_attachment_dir(repo_path)?;
-        let store_exists = Self::beads_store_footprint_exists(&beads_dir);
-
-        self.ensure_dolt_server_running(repo_path)?;
-
-        let (is_ready, reason) = if store_exists {
-            self.verify_repo_initialized(repo_path, &beads_dir)?
-        } else {
-            (false, "bd init failed".to_string())
-        };
-
-        if self.is_repo_cached_initialized(&repo_key)? && store_exists && is_ready {
-            return Ok(());
-        }
-
-        if !is_ready {
-            if store_exists {
-                self.ensure_existing_store_is_ready(repo_path, &beads_dir, &reason)?;
-            } else {
-                self.ensure_new_store_is_ready(repo_path, &beads_dir, &reason)?;
-            }
-        }
-
-        self.mark_repo_initialized(&repo_key)?;
-        Ok(())
+        self.lifecycle.ensure_repo_initialized(repo_path)
     }
 
     pub(super) fn list_tasks_impl(&self, repo_path: &Path) -> Result<Vec<TaskCard>> {
@@ -415,9 +233,6 @@ impl BeadsTaskStore {
         }
 
         if let Some(status) = patch.status {
-            if Self::task_status_requires_custom_configuration(&status) {
-                self.ensure_custom_statuses(repo_path)?;
-            }
             args.push("--status".to_string());
             args.push(status.as_cli_value().to_string());
         }
